making the crawl / parse functions more testable
